+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Create the Vivado project source
 - ADD `.gitignore`
 - Update ISA in `README`
+- Add `rars2coe.cpp` to convert rars output assemble code to coe file
 
 ## [0.0.1] - 2023-04-12